+++ nptl/sysdeps/unix/sysv/linux/s390/bits/pthreadtypes.h
@@ -77,7 +77,7 @@ typedef union
 typedef union
 {
   char __size[__SIZEOF_PTHREAD_MUTEXATTR_T];
-  long int __align;
+  int __align;
 } pthread_mutexattr_t;
 
 
@@ -100,7 +100,7 @@ typedef union
 typedef union
 {
   char __size[__SIZEOF_PTHREAD_CONDATTR_T];
-  long int __align;
+  int __align;
 } pthread_condattr_t;
